Building on mac

+++ src/mesh_control.c
@@ -2,6 +2,7 @@
 #include "mesh_control.h"
 #include "leds.h"
 #include "transport_control.h"
+#include "assert.h"
 
 static mesh_control_t m_config;
 
@@ -23,7 +24,7 @@ void mesh_control_update_config(mesh_control_t *new_config) {
 
   // Sanity checks
   if (new_config->wake_interval == 0 || new_config->wake_interval > 1000) {
-    log("Bad config: new_config->wake_interval=%d out of acceptable range "
+    logf("Bad config: new_config->wake_interval=%d out of acceptable range "
         "(1-1000)",
         new_config->wake_interval);
     return;
